% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_SID}
\alias{dat_SID}
\title{Simulated Trial Data}
\format{
## `dat_SID`
A list of data frame with both the wide format and the long format of the data:
\describe{
\item{patient_ID}{unique patient identifier}
\item{age}{patient age at entry}
\item{pre_diarrhea}{whether the patient had preceding diarrhea before entry}
\item{treat_ITT}{treatment assignment under intention-to-treat}
\item{GBS_DS}{Guillain-Barré syndrome disability scale at week 0, 1, 2, 4, 8, 12, 26}
}
}
\source{
Second intravenous immunoglobulin dose in patients with Guillain-Barré syndrome with poor prognosis (SID-GBS): a double-blind, randomised, placebo-controlled trial
Walgaard, ChristaWalgaard, Christa et al. The Lancet Neurology
}
\usage{
dat_SID
}
\description{
Simulated individual trial data based on the SID-GBS trial
}
\keyword{datasets}
