% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_corMatrix.R
\name{gen_corMatrix}
\alias{gen_corMatrix}
\title{Generate correlation matrix}
\usage{
gen_corMatrix(n_visits, rho, corstr)
}
\arguments{
\item{n_visits}{Numerical value specifying the number of visits of the repeated measurements, which
determines the dimension of the correlation matrix}

\item{rho}{Correlation coefficient}

\item{corstr}{A character string specifying the correlation structure. Options include `"independence"|"ind"`, `"exchangeable"|"cs"|"compound symmetry"|"exch"`, `"ar1"`}
}
\value{
A matrix of dimension n_visit * n_visit
}
\description{
Generate correlation matrix
}
\examples{
gen_corMatrix(n_visits=3,rho=0.6,corstr="ar1")
}
