% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwo.R
\name{lwo}
\alias{lwo}
\title{Longitudinal Win Odds}
\usage{
lwo(
  formula,
  data = parent.frame(),
  family = binomial(),
  id.var,
  visit.var,
  time.vars = NULL,
  larger = TRUE,
  corstr = "independence",
  zcor = NULL,
  std.err = "san.se",
  weights,
  contrasts = NULL
)
}
\arguments{
\item{formula}{A model formula specifying the outcome and the covariates.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{Family of the link function, only logit link gives the win odds.}

\item{id.var}{A character string specifying the cluster id. For example, patient id for repeated measurements on the same patient
Data are assumed to be sorted so that observations on each cluster appear as continuous rows in data. If data is not sorted this way,
the function will not identify the clusters correctly.}

\item{visit.var}{A character string specifying the visit within each cluster. This is needed because subjects forming a pair must come from the same visit.}

\item{time.vars}{Optional, needed if time-varying effects are modeled.
It should be a scalar or vector of variable names that represent time in the formula. This is IMPORTANT because
time variable will not be converted into pair-level.}

\item{larger}{Logical. If `TRUE` (default), larger values of the outcome indicate better outcomes.}

\item{corstr}{Working correlation structure; passed to `geepack::geese.fit`.
Options include `"independence"`, `"exchangeable"`, `"ar1"`, `"unstructured"`,
`"userdefined"`, and `"fixed"`.}

\item{zcor}{A design matrix for correlation parameters.}

\item{std.err}{Type of standard error to compute.
`"san.se"` for standard sandwich estimator
`"san.se.modified"` for the modified version, see Mancle & DeRouen, 2001, Biometrics}

\item{weights}{Optional observation-level weights.}

\item{contrasts}{A list giving contrasts for some or all of the factors appearing in the model formula. The elements of the list should have the same name as the variable and should be either a contrast matrix (specifically, any full-rank matrix with as many rows as there are levels in the factor), or else a function to compute such a matrix given the number of levels.}
}
\value{
An object of class \code{"lwo"}, inheriting from
\code{"geeglm"}, \code{"gee"}, \code{"glm"}, and \code{"lm"}.
Components include:
\itemize{
  \item \code{coefficients} – estimated regression coefficients
  \item \code{varcov} – robust covariance matrix estimate
  \item \code{fitted.values} – model-fitted probabilities
  \item \code{model.matrix} – design matrix at the pair level
  \item \code{data.pair} – transformed pair-level data
}
}
\description{
Fits a longitudinal probabilistic index model and estimate longitudinal win odds
for ordinal repeated measurements.
}
\details{
The function internally converts individual-level
data to pair-level and performs a logistic regression on the binary pair-level
outcome and covariates, then uses a sandwich estimator to obtain the correct
variance-covariance matrix accounting for pairwise correlation. Setting std.err="san.se"
gives the sandwich estimate proposed in the paper. Setting std.err = "san.se.modified" gives
a modified sandwich estimate based on the modification of Mancle & DeRouen, 2001, Biometrics.
}
\examples{
data("dat_SID")
dat_SID_long <- dat_SID$long_format
mod.lwo.spline <- lwo(
  GBS_DS ~ treat_ITT + age + pre_diarrhea + GBS_DS_baseline+
   treat_ITT:splines::ns(week,knots = 2),
 data = dat_SID_long,
 id.var = "patient_ID",
 visit.var = "week",
 time.vars = "week",
 corstr = "ar1",
 larger = FALSE
)
}
