% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pairs.R
\name{make_pairs}
\alias{make_pairs}
\title{Convert individual-level data to pair-level data}
\usage{
make_pairs(
  data,
  id.var,
  visit.var,
  outcome.var = NULL,
  covariates,
  time.vars = NULL,
  larger = TRUE
)
}
\arguments{
\item{data}{A data frame of individual-level data to be converted}

\item{id.var}{A character string specifying the cluster id. For example, patient id for repeated measurements on the same patient
Data are assumed to be sorted so that observations on each cluster appear as continuous rows in data. If data is not sorted this way,
the function will not identify the clusters correctly.}

\item{visit.var}{A character string specifying the visit within each cluster. This is needed because subjects forming a pair must come from the same visit.}

\item{outcome.var}{Optional, a character string specifying the outcome variable. Pair-level outcome values are either a `win` (1), a `loss` (0), or a `tie` (0.5) from pairwise comparisons.}

\item{covariates}{A vector of character strings specifying the covariates. Pair-level covariate values are the difference between individual-level covariate values within the pairs.}

\item{time.vars}{Optional, needed if time-varying effects are modeled.
It should be a scalar or vector of variable names that represent time in the formula. This is IMPORTANT because
time variable will not be converted into pair-level.}

\item{larger}{larger Logical. If `TRUE` (default), larger values of the outcome indicate better outcomes.}
}
\value{
A data frame containing the following columns
\describe{
  \item{L, R}{Identifiers of the left and right subjects forming each pair.}
  \item{pair.id}{Unique identifier for each subject pair.}
  \item{covariates}{Pair-level covariate differences (R - L).}
  \item{time.vars}{Time variables carried over from the left subject.}
  \item{outcome.var}{Pair-level win/loss/tie indicator (if \code{outcome.var} was supplied).}
  \item{visit.var}{Visit or time variable corresponding to each observation.}
}
}
\description{
Converts a dataset of individual-level observations into a pair-level dataset suitable for
win/loss/tie comparisons between subjects. Each row in the output corresponds to a unique subject pair
observed at a common visit.
}
\examples{
data("dat_SID")
dat_SID_long <- dat_SID$long_format
data_pair <- make_pairs(data=dat_SID_long,
id.var = "patient_ID",
visit.var = "week",
time.vars = "week",
outcome.var = "GBS_DS",
covariates = c("age","treat_ITT","pre_diarrhea","GBS_DS_baseline"),
larger = FALSE)

}
